+++ lib/togls/errors.rb
@@ -1,0 +1,3 @@
+module Togls
+  class NoFeaturesError < StandardError; end
+end